Issue-ID: I-001 Refactoring
