+++ lib/classes/category.class.dart
@@ -1,18 +1,8 @@
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
+import 'expenses.class.dart';
 
 part 'category.class.g.dart';
-
-@HiveType(typeId: 3)
-class Expense
-{
-  @HiveField(0)
-  String expenseDetails = "";
-  @HiveField(1)
-  int expense = 0;
-
-  Expense(this.expenseDetails, this.expense);
-}
 
 @HiveType(typeId: 1)
 class Category {
