+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   hive_flutter: ^1.1.0
   excel: ^2.1.0
   permission_handler: ^10.4.1
+  restart_app: ^1.2.1
 
 dev_dependencies:
   build_runner: ^2.0.0
